common/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pymongo==4.6.0
motor==3.3.1
httpx==0.25.1
python-dotenv==1.0.0
pyjwt==2.8.0
retry==0.9.2 